Fixing integration tests
