+++ services/download/src/test/resources/application.yaml
@@ -4,10 +4,11 @@ spring:
     banner-mode: off
   flyway:
     enabled: true
-    locations: classpath:/db/migration
+    locations: classpath:/db/migration, classpath:/db/specific/{vendor}
+    mixed: true
   datasource:
     driver-class-name: org.testcontainers.jdbc.ContainerDatabaseDriver
-    url: jdbc:tc:postgresql:11.8:///databasename?TC_TMPFS=/testtmpfs:rw
+    url: jdbc:tc:postgresql:11.8:///databasename?TC_TMPFS=/testtmpfs:rw?TC_INITSCRIPT=file:src/test/java/V0__init_db.sql
   test:
     database:
       # Use datasource as defined above.
